refactor: `CardCreation` -> `CardEditor` 이름 변경 및 UI 요소 추가 (#16)
- `CardCreation` 관련 클래스 및 파일 이름을 `CardEditor`로 변경
- 카드 꾸미기 화면에 뒤로가기, 삭제, 조정 버튼 UI 추가
- 버튼 UI에 필요한 리소스(`drawable`, `string`) 추가
