+++ feature/card/src/main/res/drawable/bg_circle_white_70.xml
@@ -1,0 +1,4 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid android:color="#B3FFFFFF" />
+</shape>
