+++ feature/card/src/main/res/drawable/bg_rounded_rect_white.xml
@@ -1,0 +1,5 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#FFFFFFFF" />
+    <corners android:radius="100dp" />
+</shape>
