+++ feature/card/src/main/res/drawable/ic_trash_bin.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="22dp"
+    android:height="22dp"
+    android:viewportWidth="22"
+    android:viewportHeight="22">
+  <path
+      android:pathData="M2.75,5.5H4.583M4.583,5.5H19.25M4.583,5.5V18.333C4.583,18.82 4.776,19.286 5.12,19.63C5.464,19.974 5.93,20.167 6.417,20.167H15.583C16.07,20.167 16.536,19.974 16.88,19.63C17.223,19.286 17.417,18.82 17.417,18.333V5.5M7.333,5.5V3.667C7.333,3.18 7.526,2.714 7.87,2.37C8.214,2.027 8.68,1.833 9.167,1.833H12.833C13.32,1.833 13.786,2.027 14.13,2.37C14.474,2.714 14.667,3.18 14.667,3.667V5.5M9.167,10.083V15.583M12.833,10.083V15.583"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#1E1E1E"
+      android:strokeLineCap="round"/>
+</vector>
