+++ feature/card/src/main/res/layout/activity_card_editor.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.example.card.ui.cardeditor.CardEditorActivity">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/unity_container"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageButton
+        android:id="@+id/img_btn_back"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
+        android:background="@drawable/bg_circle_white_70"
+        android:layout_marginTop="30dp"
+        android:layout_marginStart="20dp"
+        android:contentDescription="@string/creation_cd_back_img_button"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:srcCompat="@drawable/ic_arrow_left" />
+
+    <ImageButton
+        android:id="@+id/img_btn_delete"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
+        android:background="@drawable/bg_circle_white"
+        android:elevation="2dp"
+        android:layout_marginBottom="20dp"
+        android:layout_marginStart="20dp"
+        android:contentDescription="@string/creation_cd_delete_img_button"
+        app:layout_constraintBottom_toBottomOf="@+id/unity_container"
+        app:layout_constraintStart_toStartOf="parent"
+        app:srcCompat="@drawable/ic_trash_bin" />
+
+    <android.widget.Button
+        android:id="@+id/btn_adjust"
+        android:layout_width="wrap_content"
+        android:layout_height="48dp"
+        android:text="@string/creation_btn_adjust"
+        android:textAllCaps="false"
+        android:fontFamily="@font/irish_grover_regular"
+        android:backgroundTint="#FFFFFF"
+        android:background="@drawable/bg_rounded_rect_white"
+        android:layout_marginBottom="20dp"
+        android:layout_marginEnd="20dp"
+        app:layout_constraintBottom_toBottomOf="@+id/unity_container"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+    <androidx.compose.ui.platform.ComposeView
+        android:id="@+id/compose_container"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/unity_container"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
