+++ feature/card/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="creation_cd_back_img_button">Go back</string>
+    <string name="creation_cd_delete_img_button">Delete Element</string>
+
+    <string name="creation_btn_adjust">Adjust</string>
+</resources>
